feat: 시간 관리 페이지 API 기능 구현
