+++ README.md
@@ -9,7 +9,7 @@
   - [x] 예약 조회
   - [x] 예약 추가
   - [x] 예약 취소
-- [ ] 시간관리 기능
+- [x] 시간관리 기능
   - [x] 시간 추가 API 기능 구현
   - [x] 시간 조회 API 기능 구현
-  - [ ] 시간 삭제 API 기능 구현
+  - [x] 시간 삭제 API 기능 구현
