+++ src/test/java/roomescape/AdminTest.java
@@ -1,0 +1,32 @@
+package roomescape;
+
+import static org.hamcrest.Matchers.is;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+class AdminTest {
+
+    @DisplayName("어드민 메인 페이지 테스트")
+    @Test
+    void adminMainPage() {
+        RestAssured.given().log().all()
+                .when().get("/admin")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("예약 어드민 페이지 테스트")
+    @Test
+    void readReservations() {
+        RestAssured.given().log().all()
+                .when().get("/admin/reservation")
+                .then().log().all()
+                .statusCode(200);
+    }
+}
